pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-oracle"
version = "0.0.2"
description = "RAG Oracle – failure diagnostics for Retrieval-Augmented Generation systems"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MUHAMMAD HASSAN SHAHBAZ"}
]
keywords = ["rag", "llm", "retrieval-augmented-generation", "failure-detection", "diagnostics", "root-cause-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-text-splitters>=0.0.1",
    "langchain-groq>=0.0.1",
    "pypdf>=3.17.0",
    "chromadb>=0.4.22",
    "langchain-huggingface>=0.0.1",
    "sentence-transformers>=2.2.2",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/Rana-Hassan7272/RAG-ORACLE"
Documentation = "https://github.com/Rana-Hassan7272/RAG-ORACLE#readme"
Repository = "https://github.com/Rana-Hassan7272/RAG-ORACLE"
Issues = "https://github.com/Rana-Hassan7272/RAG-ORACLE/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag_pipeline*"]

[tool.setuptools.package-data]
rag_pipeline = ["*.md"]

